# fix(federation): several reuse_fragments() performance improvements