… apollo-federation/src/query_plan/fetch_dependency_graph.rs
… apollo-federation/src/schema/definitions.rs
… apollo-federation/src/schema/position.rs