This does *not* fix our trouble yet, but it addresses a difference in JS and Rust implementations that made Rust substantially slower. After these changes the `FieldsInSetCanMerge` validation doesn't dominate time spent on the queries I'm using.

Also added a cache for `FragmentRestrictionsAtType`, just like the JS has. Plus some miscellaneous smaller improvements.

Finally, renamed the extremely generic `optimize()` to `reuse_fragments()` :)

With one example query:
| Before | After |
|-|-|
| 14.91s | 5.12s |